requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
aiofiles==23.2.1
requests==2.31.0
python-magic==0.4.27
ffmpeg-python==0.2.0
asyncio==3.4.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1